Created Add Workout back end
Main function was to create the back end portion of the Add Workout. To do so, had to create  an additional table in the database(workout_plans_exercises). Also created the WorkoutPlan.ts in order to add the SQLIte insert and select functionality from the db for exercises. This commit also allows one to see their added exercises from the database on the WorkoutDay page. Added an extra detail in ExerciseDetail page.